privs and pvc: No more PVCs needed with a combined task
While this could now be trimmed to just be a task and taskrun workflow,
using a pipeline allows us to keep the task more similar to the two
upstream tasks and control the inputs a bit more.

Since this allows the steps of the task to all share one emptydir volume,
you no longer need a pvc, and you are sure to have it cleaned up after
the build is done.