deploy: Patch the correct images in the correct places
As we are not testing the validation webhooks locally, we missed that
the images were being pulled from upstream, and not from our image
repositories.

This modifies the patches to patch in the correct place, at the
correct time.

Note that this binds all the image tags for all the environments (as
is), so we might want to change it in the future, but probably will
move to helm/helmfile before that so will have to solve it then.

Bug: T316854